+++ src/Protocol/Parser/ProtocolParser.php
@@ -17,6 +17,7 @@ use Dorpmaster\Nats\Protocol\MsgMessage;
 use Dorpmaster\Nats\Protocol\NatsMessageType;
 use Dorpmaster\Nats\Protocol\OkMessage;
 use Dorpmaster\Nats\Protocol\PingMessage;
+use Dorpmaster\Nats\Protocol\PongMessage;
 
 final readonly class ProtocolParser implements ProtocolParserInterface
 {
@@ -56,6 +57,7 @@ final readonly class ProtocolParser implements ProtocolParserInterface
                 NatsMessageType::MSG => yield from self::parseToMsg($metadata),
                 NatsMessageType::HMSG => yield from self::parseToHMsg($metadata),
                 NatsMessageType::PING => new PingMessage(),
+                NatsMessageType::PONG => new PongMessage(),
                 NatsMessageType::OK => new OkMessage(),
                 NatsMessageType::ERR => new ErrMessage(implode(' ', $metadata)),
                 default => throw new \RuntimeException(sprintf('Unknown message type "%s"', $type)),